feat: caso de uso de adicionar pessoa e de buscar ordenada

+++ fila_core/src/main/java/com/vinicius123131/entidades/Pessoa.java
@@ -12,5 +12,5 @@ public abstract class Pessoa implements PessoaNaFila {
     protected Long id;
     protected String nome;
     protected Integer idade;
-    protected Long posicao_fila;
+    protected Integer posicao_fila;
 }
